+++ partners_site/shop/models.py
@@ -93,7 +93,7 @@ class ProductGroup(models.Model):
 # Create your models here.
 class Product(models.Model):
     name = models.CharField(max_length=255, verbose_name='Наименование')
-    amo_id = models.PositiveIntegerField(verbose_name='id товара в amocrm')
+    amo_id = models.PositiveBigIntegerField(verbose_name='id товара в amocrm')
     price = models.PositiveIntegerField(verbose_name='Цена')
     title = models.TextField(verbose_name='Описание')
     short_description = models.CharField(max_length=200, blank=True, verbose_name='Краткое описание')
